BRMS-423: Update Payload validation to include reference validation
